Close clippy/CI gap: add clippy to CI + fix pre-existing lint drift
CI ran llvm-cov + audits + vitest but never clippy, while reverify.toml requires
`clippy -D warnings` — so lint drift accumulated invisibly and made `keel reverify`
red, blocking deliverableTestGate. (issueClippyNotInCi)

- Guard (D0047): new server-clippy CI job running the exact reverify clippy command.
- Fixed the ~7 pre-existing lints in unrelated files (redundant cast, map_or→
  is_some_and, ==false→!, too_many_arguments allow, items_after_test_module allows
  for the ClientIp/AuthToken extractors, OpenLocks type alias, repeat().take()→vec!).

clippy --all-targets -D warnings exits 0; 99 server tests green. Resolves
issueClippyNotInCi via clippyGateInCi.

+++ server/src/audit.rs
@@ -1,3 +1,6 @@
+// The ClientIp request-extractor impl intentionally sits at the file's end (after the test module);
+// suppress clippy's items-after-test-module STYLE lint rather than reorder the file.
+#![allow(clippy::items_after_test_module)]
 // SEC-CMMC (AU, NIST SP 800-171 3.3.x): a dedicated, structured, attributable SECURITY-AUDIT trail.
 // Before this, security events were scattered as free-text lines in the operational stdout stream,
 // untimestamped and without a source, and 8 event classes emitted nothing at all. This module is the
@@ -86,7 +89,7 @@ fn rfc3339_from_epoch(secs: u64, millis: u32) -> String {
     // civil_from_days: days since 1970-01-01 -> (year, month, day)
     let z = days + 719_468;
     let era = (if z >= 0 { z } else { z - 146_096 }) / 146_097;
-    let doe = (z - era * 146_097) as i64; // [0, 146096]
+    let doe = z - era * 146_097; // [0, 146096] (already i64)
     let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365; // [0, 399]
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
